refactor(admin): extract AgentContext from domain/trace.rs into domain/agent_context.rs (PIP-687)
domain/trace.rs was 1506 lines, 6 over the 1500-line hard limit. Move
AgentContext + AgentContextTrust + all their helpers and header-merge
logic into a dedicated module, keeping the public re-export path
unchanged (crate::gateway::admin::trace::*).

- New: domain/agent_context.rs (941 lines)
- trace.rs: 615 lines (was 1506) — well under the 1500 cap
- Backward-compat: all types still available through domain/trace.rs
